+++ docs/CHANGELOG.md
@@ -96,6 +96,8 @@ Semantic Versioning.
     that the inferred variable was redeclared.
   * `class implements I {}` is now parsed as a class with no name rather than a
     class with the name `implements`.
+  * `class C extends Base<T> implements I {}` now correctly parses `Base<T>` as
+    a type with generic arguments rather than logical comparisons.
   * Variables declared inside a `declare global` block are now correctly
     declared as global variables instead of module variables. This means that
     variables inside a `declare global` block can be shadowed by module